### 2018-09-14 279d55b
replace R2 with math equations
