### 2018-09-14 c54617f
corrected typos
