+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @article{ives2018r2s,
   title={R2s for correlated data: Phylogenetic models, LMMs, and GLMMs},
-  author={Ives, Anthony},
+  author={Ives, Anthony R},
   journal={Systematic Biology},
   pages={--},
   year={In press}
