### 2018-10-18 #1028 @whedon
[REVIEW]: `rr2`: An R package to calculate R2s for regression models
**Submitting author:** @arives (<a href="http://orcid.org/0000-0001-9375-9523">Anthony Ives</a>)
**Repository:** <a href="https://github.com/arives/rr2" target ="_blank">https://github.com/arives/rr2</a>
**Version:** v1.0.0
