### 2018-10-26 8f74c93
fixed R2_resid and R2_pred
In R2_resid, changed formula for option sigma2_d = 's2w' for binomial and poisson models. Made this the default option.

In R2.pred, fixed the code for glm and glmm to allow binomial distributions with size > 1.
