### 2018-10-26 arives/rr2#5 @sastoudt
Licence
For JOSS Review: https://github.com/openjournals/joss-reviews/issues/1028

- [ ] In LICENSE.md finish at "END OF TERMS AND CONDITIONS"
- [ ] "add information on how to contact you by electronic and paper mail" in the headers of the source files
