### 2018-10-26 arives/rr2#6 @sastoudt
References
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

- [ ] DOI for Ecological Monographs paper
- [ ] DOI for Systematic Biology paper
- [ ] update In press for Systematic Biology now that published
