### 2018-10-26 arives/rr2#7 @sastoudt
Community guidelines
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

- [ ] add community guidelines to README.md for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
