### 2018-10-26 arives/rr2#8 @sastoudt
Typos
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

- [ ] in README.md Package structure section, second sentence: reduce model --> reduced model
- [ ] in test_r2s.R lines 61 and 85 creat --> create
