### 2018-10-26 arives/rr2#9 @sastoudt
Example Usage
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

- [x] In the JOSS paper you mention that this package can be used to "compare contributions of predictor variables and/or correlation structure". It would be nice to have an example of this in the README.md
- [x] add example usage of partialR2() and partialR2adj() to README.md
- [ ] add example usage of communityPGLMM() to README.md
- [x] note change in binaryPGLMM() from the ape package in README.md
- [x]  add example usage of different values for parameter sigma2_d in README.md
