### 2018-10-26 arives/rr2#10 @sastoudt
Version
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

Version in DESCRIPTION (0.3.0) does not match the GitHub release (v1.0.0)
