### 2018-10-26 arives/rr2#11 @sastoudt
Tests
For JOSS review: https://github.com/openjournals/joss-reviews/issues/1028

- [x] testthat.R results in Error: No tests found for rr2

I ran the tests manually, and they passed. However, for the commented out tests I was not able to test them manually because phyr is not available for R version 3.5.1, 3.3.3, 3.2.5, or 3.1.3.

- [x] include expect_message tests regarding lack of particular methods for communityPGLM
