### 2018-10-27 arives/rr2#5 @daijiang
Hi @sastoudt,
For the first point, you mean we should remove all things after "END OF TERMS AND CONDITIONS" in LICENSE.md?
What do you mean by the second point? I think the contact information is already in the DESCRIPTION file and that should be enough?
