### 2018-10-27 arives/rr2#10 @daijiang
Updated the DESCRIPTION file to v1.0.0.
