### 2018-10-27 26bd62a
update version and citation in DESCRIPTION; close issue #10
