### 2018-10-27 arives/rr2#11 @daijiang
Hi @sastoudt
Thanks for your help with the package! I think the Error is because you tried to run `test that.R`.  If you use `devtools::test(pkg = ".")` to test the package, you should get no error.

We commented out the test about `phyr` because this package is still under development and is not stable yet. We added expect message about communityPGLMM. We will uncomment these sections when `phyr` is stable.
