### 2018-10-27 c4edc29
add expect message for communityPGLMM #11 close #8
