### 2018-10-27 3860c20
update citations with doi close #6
