+++ joss/paper.bib
@@ -1,9 +1,14 @@
 @article{ives2018r2s,
-  title={R2s for correlated data: Phylogenetic models, LMMs, and GLMMs},
-  author={Ives, Anthony R},
-  journal={Systematic Biology},
-  pages={--},
-  year={In press}
+author = {Ives, Anthony R},
+title = {R$^{2}$s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs},
+journal = {Systematic Biology},
+volume = {},
+number = {},
+pages = {syy060},
+year = {2018},
+doi = {10.1093/sysbio/syy060},
+URL = {http://dx.doi.org/10.1093/sysbio/syy060},
+eprint = {/oup/backfile/content_public/journal/sysbio/pap/10.1093_sysbio_syy060/4/syy060.pdf}
 }
 
 @article{ives2011generalized,
@@ -14,7 +19,8 @@
   number={3},
   pages={511--525},
   year={2011},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1890/10-1264.1}
 }
 
 @incollection{ives2014phylogenetic,
@@ -26,14 +32,14 @@
   publisher={Springer}
 }
 
- @Article{lme4r,
-    title = {Fitting Linear Mixed-Effects Models Using {lme4}},
-    author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and
-      Steve Walker},
-    journal = {Journal of Statistical Software},
-    year = {2015},
-    volume = {67},
-    number = {1},
-    pages = {1--48},
-    doi = {10.18637/jss.v067.i01},
-  }
+@Article{lme4r,
+  title = {Fitting Linear Mixed-Effects Models Using {lme4}},
+  author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and
+    Steve Walker},
+  journal = {Journal of Statistical Software},
+  year = {2015},
+  volume = {67},
+  number = {1},
+  pages = {1--48},
+  doi = {10.18637/jss.v067.i01},
+}
