+++ R/r2.R
@@ -37,7 +37,7 @@ NULL
 #' @param pred Whether to calculate R2.pred; default is TRUE.
 #' @return A vector, with all three R2s by default.
 #' @author Daijiang Li and Anthony R. Ives
-#' @references Ives A. in press. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology.
+#' @references Ives A. 2018. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology. <doi:10.1093/sysbio/syy060>
 #' @seealso MuMIn, lme4, ape, phylolm, pez
 #' @export
 #' @examples library(ape)
