+++ R/r2_pred.R
@@ -33,7 +33,7 @@
 #' For compatibility and generating reduced models, rr2 will compute R2.pred for LM and GLM that correspond to LMM/PGLS and GLMM/PGLMM.
 #' 
 #' @author Anthony R. Ives
-#' @references Ives A. in press. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology.
+#' @references Ives A. 2018. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology. <doi:10.1093/sysbio/syy060>
 #' @seealso MuMIn, lme4, ape, phylolm, pez
 #' 
 #' @examples library(ape)
