+++ man/R2.Rd
@@ -155,7 +155,7 @@ R2(z.f, phy = phy)
 
 }
 \references{
-Ives A. in press. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology.
+Ives A. 2018. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology. <doi:10.1093/sysbio/syy060>
 }
 \seealso{
 MuMIn, lme4, ape, phylolm, pez
