### 2018-10-27 3883e59
add community guide, fix typos, close #7
