### 2018-10-27 fa518a0
complete sentence
