### 2018-10-27 b91ea8d
update rr2 then knit readme
