+++ README.md
@@ -136,37 +136,29 @@ z.0.lmm <- lm(y_re_intercept ~ 1, data = d)
 R2(mod = z.f.lmm, mod.r = z.x.lmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.5356524
-    ## 2 R2_resid 0.6036311
-    ## 3  R2_pred 0.6087728
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.5356524 0.6036311 0.6087728
 
 ``` r
 R2(mod = z.f.lmm, mod.r = z.v.lmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.7441745
-    ## 2 R2_resid 0.8373347
-    ## 3  R2_pred 0.8559029
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.7441745 0.8373347 0.8559029
 
 ``` r
 R2(mod = z.f.lmm, mod.r = z.0.lmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.7762978
-    ## 2 R2_resid 0.8767789
-    ## 3  R2_pred 0.8991618
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.7762978 0.8767789 0.8991618
 
 ``` r
 R2(mod = z.f.lmm) # if omit mod.r, default will be the simplest model, such as z.0.lmm here.
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.7762978
-    ## 2 R2_resid 0.8767789
-    ## 3  R2_pred 0.8991618
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.7762978 0.8767789 0.8991618
 
 GLMM
 ----
@@ -179,28 +171,22 @@ z.v.glmm <- glm(y_binary ~ x1, data = d, family = "binomial")
 R2(mod = z.f.glmm, mod.r = z.x.glmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.1170588
-    ## 2 R2_resid 0.2254840
-    ## 3  R2_pred 0.1373521
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.1170588 0.1413694 0.1373521
 
 ``` r
 R2(mod = z.f.glmm, mod.r = z.v.glmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.1990563
-    ## 2 R2_resid 0.4246935
-    ## 3  R2_pred 0.3545240
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.1990563 0.3404476 0.3545240
 
 ``` r
 R2(mod = z.f.glmm)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.2406380
-    ## 2 R2_resid 0.4530099
-    ## 3  R2_pred 0.3792381
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.2406380 0.3659939 0.3792381
 
 PGLS
 ----
@@ -213,19 +199,15 @@ z.v.pgls <- lm(y_pgls ~ x_trait, data = d)
 R2(mod = z.f.pgls, mod.r = z.v.pgls, phy = phy)
 ```
 
-    ##        R2s     value
-    ## 1   R2_lik 0.2353912
-    ## 2 R2_resid 0.3590018
-    ## 3  R2_pred 0.3114035
+    ##    R2_lik  R2_resid   R2_pred 
+    ## 0.2353912 0.3590018 0.3114035
 
 ``` r
 R2(mod = z.f.pgls, phy = phy)
 ```
 
[…]
