### 2018-10-27 618a103
add more examples in readme #9
