### 2018-10-27 b6c924a
different sigma2_d examples in readme #9
