### 2018-10-27 arives/rr2#9 @daijiang
@sastoudt Thanks! We have added more examples in README.md. However, we did not add example for `phyr::commuintyPGLMM()` because it is not stable yet. At this stage, we don't want to put `phyr` as a package imported or suggested by `rr2`, which will be required if we add examples about it in README (and add tests). We have tested the usage of `rr2::R2()` on `phyr::commuintyPGLMM()` locally. When `phyr` is stable in the near future, we will add examples in the README.md and tests in the test folder.
