### 2018-10-28 839f15e
update readme with more examples
