### 2018-10-28 arives/rr2#5 @sastoudt
First point: Yes, looking [here](https://opensource.org/licenses/GPL-3.0) I think everything after END OF TERMS AND CONDITIONS is for the authors rather than the reader of the license.

Second point: You're right, I missed the e-mails in the DESCRIPTION.
