### 2018-10-30 fafabae
Did some cleaning
I removed some commented-out code and did a little editing of the documentation. I also retested the examples.
