+++ R/r2_lik.R
@@ -7,17 +7,21 @@
 #' @return R2.lik value.
 #' @export
 #'
-#' @details  \code{R2.lik()} works with classes 'lm', 'glm', 'lmerMod', 'glmerMod', 'phylolm', 'phyloglm', and 'communityPGLMM' (family =  'gaussian' only). 
+#' @details  \code{R2.lik()} works with classes 'lm', 'glm', 'lmerMod', 'glmerMod', 'phylolm', 'phyloglm', and 'communityPGLMM' (family =  'gaussian' only). It is implemented as
 #' 
 #' \deqn{partial R2 = 1 - exp(-2/n * (logLik(mod.f) - logLik(mod.r)))}
 #' 
-#' where mod.f and mod.r are the full and reduced models, respectively. The total R2 is given when mod.r is the model corresponding to mod.f that contains only the intercept.
+#' where 'mod.f' and 'mod.r' are the full and reduced models, respectively. The total R2 is given when 'mod.r' is the model corresponding to mod.f that contains only the intercept. For GLMMs and PGLMMs, \code{R2.lik()} is standardized to have a maximum of one following Nagelkerke (1991).
 #' 
-#' R2.lik is also computed for LMMs and GLMMs in the {MuMIn} package.
+#' Note that \code{phyloglm()} can have difficulties in finding solutions when there is no phylogenetic signal. Therefore, when the estimate of alpha is >50, indicating no phylogenetic signal, the model is refit with the corresponding GLM.
+#' 
+#' \code{R2.lik()} is also computed for LMMs and GLMMs in the {MuMIn} package.
 #' 
 #' @seealso MuMIn, lme4, ape, phylolm, pez
 #' @author Anthony R. Ives
 #' @references Ives A. 2018. R2s for Correlated Data: Phylogenetic Models, LMMs, and GLMMs. Systematic Biology. <doi:10.1093/sysbio/syy060>
+#' 
+#' Nagelkerke 1991. A note on a general definition of the coefficient of determination. Biometrika 78:691–692.
 #' 
 #' @examples library(ape)
 #' library(phylolm)
@@ -52,7 +56,7 @@
 #' R2.lik(z.f, z.v)
 #' R2.lik(z.f)
 #' 
-#' # These give the same results
+#' # These give the same results.
 #' R2.lik(z.f, z.0)
 #' R2.lik(z.f)
 #' 
@@ -148,8 +152,8 @@ R2.lik <- function(mod = NULL, mod.r = NULL) {
     
     if (class(mod)[1] == "lm") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- lm(Y ~ 1)
+            y <- model.frame(mod)[, 1]
+            mod.r <- lm(y ~ 1)
         }
         if (!is.element(class(mod.r)[1], c("lm"))) {
             stop("mod.r must be class lm.")
@@ -159,8 +163,8 @@ R2.lik <- function(mod = NULL, mod.r = NULL) {
     
     if (class(mod)[1] == "glm") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- glm(Y ~ 1, family = family(mod)[[1]])
+            y <- model.frame(mod)[, 1]
+            mod.r <- glm(y ~ 1, family = family(mod)[[1]])
         }
         if (!is.element(class(mod.r)[1], c("glm"))) {
             stop("mod.r must be class glm.")
@@ -173,8 +177,8 @@ R2.lik <- function(mod = NULL, mod.r = NULL) {
     
     if (class(mod)[1] == "lmerMod") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
[…]
