+++ R/r2_pred.R
@@ -28,6 +28,9 @@
 #' 
 #' \deqn{partial R2 = 1 - (1 - R2.pred.f)/(1 - R2.pred.r)}
 #' 
+#' Note that \code{phylolm()} can have difficulties in finding solutions when there is no phylogenetic
+#' signal; when the estimate indicates no phylogenetic signal, you should refit the model with the corresponding LM.
+#' 
 #' \strong{LM (lm) and GLM (glm):} 
 #' 
 #' For compatibility and generating reduced models, rr2 will compute R2.pred for LM and GLM that correspond to LMM/PGLS and GLMM/PGLMM.
@@ -162,8 +165,8 @@ R2.pred <- function(mod = NULL, mod.r = NULL, phy = NULL) {
     
     if (class(mod)[1] == "lm") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- lm(Y ~ 1)
+            y <- model.frame(mod)[, 1]
+            mod.r <- lm(y ~ 1)
         }
         if (!is.element(class(mod.r)[1], c("lm"))) {
             stop("mod.r must be class lm.")
@@ -173,8 +176,8 @@ R2.pred <- function(mod = NULL, mod.r = NULL, phy = NULL) {
     
     if (class(mod)[1] == "glm") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- glm(Y ~ 1, family = family(mod)[[1]])
+            y <- model.frame(mod)[, 1]
+            mod.r <- glm(y ~ 1, family = family(mod)[[1]])
         }
         if (!is.element(class(mod.r)[1], c("glm"))) {
             stop("mod.r must be class glm.")
@@ -187,8 +190,8 @@ R2.pred <- function(mod = NULL, mod.r = NULL, phy = NULL) {
     
     if (class(mod)[1] == "lmerMod") {
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- lm(Y ~ 1)
+            y <- model.frame(mod)[, 1]
+            mod.r <- lm(y ~ 1)
         }
         if (class(mod.r)[1] == "merModLmerTest") 
             class(mod.r) <- "lmerMod"
@@ -199,11 +202,9 @@ R2.pred <- function(mod = NULL, mod.r = NULL, phy = NULL) {
     }
     
     if (class(mod)[1] == "glmerMod") {
-        # if (family(mod)[[1]] != 'binomial') { stop('Sorry, but only binomial (binary)
-        # glmerMod models are allowed.') }
         if (!is.object(mod.r)) {
-            Y <- model.frame(mod)[, 1]
-            mod.r <- glm(Y ~ 1, family = family(mod)[[1]])
+            y <- model.frame(mod)[, 1]
+            mod.r <- glm(y ~ 1, family = family(mod)[[1]])
         }
         if (!is.element(class(mod.r)[1], c("glmerMod", "glm"))) {
             stop("mod.r must be class glmerMod or glm.")
@@ -216,11 +217,11 @@ R2.pred <- function(mod = NULL, mod.r = NULL, phy = NULL) {
     
     if (class(mod)[1] == "phylolm") {
         if (!is.object(phy)) {
[…]
