### 2018-10-30 #1028 @arives
I just archived on Zenodo:

.. image:: https://zenodo.org/badge/98235954.svg
   :target: https://zenodo.org/badge/latestdoi/98235954
