### 2018-10-31 4aa414c
update citations
