+++ README.Rmd
@@ -12,7 +12,7 @@ knitr::opts_chunk$set(echo = TRUE)
 
 # Goal
 
-This package provides three R^2^s for statistical models with correlated errors including classes: 'lmerMod' (LMM), 'glmerMod' (GLMM), 'phylolm' (Phylogenetic GLS), and 'binaryPGLMM/phyloglm/communityPGLMM' (Phylogenetic Logistic Regression). Detailed technical descriptions can be found in [Ives 2017, preprint](https://www.biorxiv.org/content/early/2017/07/26/144170).
+This package provides three R^2^s for statistical models with correlated errors including classes: 'lmerMod' (LMM), 'glmerMod' (GLMM), 'phylolm' (Phylogenetic GLS), and 'binaryPGLMM/phyloglm/communityPGLMM' (Phylogenetic Logistic Regression). Detailed technical descriptions can be found in [Ives 2018](https://doi.org/10.1093/sysbio/syy060).
 
 # Installation
 
