### 2018-10-31 #1028 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
